# Conflicts:
#	src/MEGASync/MegaApplication.cpp

+++ src/MEGASync/control/Utilities.h
@@ -5,6 +5,7 @@
 #include <QHash>
 #include <QPixmap>
 #include <QDir>
+#include <QIcon>
 
 #include <sys/stat.h>
 
@@ -19,8 +20,8 @@ struct PlanInfo
 {
     int amount;
     QString currency;
-    unsigned long long gbStorage;
-    unsigned long long gbTransfer;
+    int gbStorage;
+    int gbTransfer;
     int level;
 };
